chore: norm main, split up init
