+++ src/main.c
@@ -3,56 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tmeniga@student.42vienna.com <tmeniga>     +#+  +:+       +#+        */
+/*   By: aprevrha <aprevrha@student.42vienna.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/14 22:34:46 by aprevrha          #+#    #+#             */
-/*   Updated: 2024/11/27 19:48:27 by tmeniga@stu      ###   ########.fr       */
+/*   Updated: 2024/11/30 19:22:50 by aprevrha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/cub3d.h"
-
-
-void	free_and_exit(t_data *data, int code)
-{
-	if (data->players)
-		free(data->players);
-	get_next_line(data->fd, 1);
-	if (data->map && data->map->texture_data)
-	{
-		if (data->map->texture_data->textures[0].img)
-			mlx_destroy_image(data->mlx,
-				data->map->texture_data->textures[0].img);
-		if (data->map->texture_data->textures[1].img)
-			mlx_destroy_image(data->mlx,
-				data->map->texture_data->textures[1].img);
-		if (data->map->texture_data->textures[2].img)
-			mlx_destroy_image(data->mlx,
-				data->map->texture_data->textures[2].img);
-		if (data->map->texture_data->textures[3].img)
-			mlx_destroy_image(data->mlx,
-				data->map->texture_data->textures[3].img);
-		free_texture_data(data->map->texture_data);
-	}
-	if (data->map)
-		free_map(data->map);
-	if (data->img)
-		mlx_destroy_image(data->mlx, data->img);
-	if (data->win)
-		mlx_destroy_window(data->mlx, data->win);
-	if (data->mlx)
-	{
-		mlx_destroy_display(data->mlx);
-		free(data->mlx);
-	}
-	exit(code);
-}
-
-int	win_close_button(t_data *data)
-{
-	free_and_exit(data, 0);
-	return (0);
-}
 
 int	is_file_extension(char *filename, char *extension)
 {
@@ -81,44 +39,24 @@ int	load_texture(char *path, t_data *data, t_texture *texture)
 	return (0);
 }
 
-void	init2null(t_data *data)
+int	init_textures(t_data *data)
 {
-	struct timeval	tv_now;
-	data->mlx = NULL;
-	data->win = NULL;
-	data->img = NULL;
-	data->addr = NULL;
-	data->map = NULL;
-	data->players = NULL;
-	data->use_mouse = false;
-	if (gettimeofday(&tv_now, NULL))
-		printf("gettimeofday failed");
-	data->lastframe = tv_now;
-	// data->map->texture_data = NULL;
-	// data->map->texture_data->path_no = NULL;
-	// data->map->texture_data->path_so = NULL;
-	// data->map->texture_data->path_we = NULL;
-	// data->map->texture_data->path_ea = NULL;
-	// data->map->arr = NULL;
-}
-
-int		init_textures(t_data *data)
-{	
 	data->map->texture_data->textures[0].img = NULL;
 	data->map->texture_data->textures[1].img = NULL;
 	data->map->texture_data->textures[2].img = NULL;
 	data->map->texture_data->textures[3].img = NULL;
-
-	
-	if (load_texture(data->map->texture_data->path_no, data, &data->map->texture_data->textures[0]))
+	if (load_texture(data->map->texture_data->path_no, data,
+			&data->map->texture_data->textures[0]))
 		return (printf("Error: could not load texture\n"), 1);
-	if (load_texture(data->map->texture_data->path_ea, data, &data->map->texture_data->textures[1]))
+	if (load_texture(data->map->texture_data->path_ea, data,
+			&data->map->texture_data->textures[1]))
 		return (printf("Error: could not load texture\n"), 1);
-	if (load_texture(data->map->texture_data->path_so, data, &data->map->texture_data->textures[2]))
+	if (load_texture(data->map->texture_data->path_so, data,
+			&data->map->texture_data->textures[2]))
 		return (printf("Error: could not load texture\n"), 1);
-	if (load_texture(data->map->texture_data->path_we, data, &data->map->texture_data->textures[3]))
+	if (load_texture(data->map->texture_data->path_we, data,
+			&data->map->texture_data->textures[3]))
 		return (printf("Error: could not load texture\n"), 1);
-
 	return (0);
 }
 
@@ -134,43 +72,24 @@ int	check_input(int argc, char **argv)
 int	main(int argc, char **argv)
 {
 	t_data	data;
-	int		fd;
-	
+
 	init2null(&data);
 	if (check_input(argc, argv))
 		return (1);
-	
-	fd = open(argv[1], O_RDONLY);
-	if (fd < 0)
+	data.fd = open(argv[1], O_RDONLY);
+	if (data.fd < 0)
 		return (1);
-	data.fd = fd;
-
-	
-	if ((data.map = parse_map(fd)) == NULL)
+	data.map = parse_map(data.fd);
+	if (data.map == NULL)
 		free_and_exit(&data, 1);
-				
 	if (init_players(&data.players, data.map))
 		free_and_exit(&data, 1);
-	
 	if (init_mlx(&data))
 		free_and_exit(&data, 1);
-	
 	if (init_textures(&data))
 		free_and_exit(&data, 1);
-	render_map(&data);
-	mlx_put_image_to_window(data.mlx, data.win, data.img, 0, 0);
 	init_keys(data.keys, data.players);
-	mlx_hook(data.win, 2, KeyPressMask, handle_keydown, data.keys);
-	mlx_hook(data.win, 3, KeyReleaseMask, handle_keyup, data.keys);
-	mlx_hook(data.win, 17, StructureNotifyMask, win_close_button, &data);
-	mlx_hook(data.win, MotionNotify, PointerMotionMask, handle_mousemove,
-		&data);
-	mlx_mouse_hook(data.win, handle_mouseclick, &data);
-	// test texture
-	// printtexture_data(*data.map->texture_data);
-
-	mlx_loop_hook(data.mlx, loop_hook, &data);
+	init_hooks(&data);
 	mlx_loop(data.mlx);
-
 	free_and_exit(&data, 0);
-}	
+}
